Moved the controllers and their test files into a seperate folder
